delete and find article by Id

+++ src/main/java/com/belvinard/gestiondestock/repositories/ArticleRepository.java
@@ -6,11 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ArticleRepository extends JpaRepository<Article, Integer> {
+public interface ArticleRepository extends JpaRepository<Article, Long> {
 
   Optional<Article> findArticleByCodeArticle(String codeArticle);
 
-  List<Article> findAllByCategoryId(Integer idCategory);
-
+  List<Article> findAllByCategoryId(Long idCategory);
 
 }
